market_summary/requirements.txt
requests>=2.31.0
yfinance>=0.2.30
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
